Pojoak eginda eta letra larriz jarrita
