+++ Bidai_Agentzia/bin/.gitignore
@@ -1,2 +1,1 @@
-/aplikazioa/
 /modeloa/
