--- Bidai_Agentzia/src/modeloa/hiria.java
@@ -1,5 +1,0 @@
-package modeloa;
-
-public class hiria {
-
-}
